image-searching/requirements.txt
torch>=2.0.0
open-clip-torch>=2.20.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
